# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AssetHierarchyResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AssetHierarchyItem"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "AssetHierarchyResponse",
        "type": "object"
      },
      "DiscrepancyAuditResponse": {
        "properties": {
          "findings": {
            "items": {
              "$ref": "#/components/schemas/DiscrepancyFindingItem"
            },
            "title": "Findings",
            "type": "array"
          },
          "observation_id": {
            "format": "uuid",
            "title": "Observation Id",
            "type": "string"
          }
        },
        "required": [
          "observation_id",
          "findings"
        ],
        "title": "DiscrepancyAuditResponse",
        "type": "object"
      },
...
      "EquipmentDetail": {
        "properties": {
          "cell_id": {
            "format": "uuid",
            "title": "Cell Id",
            "type": "string"
          },
          "equipment_class": {
            "title": "Equipment Class",
            "type": "string"
          },
          "equipment_code": {
            "title": "Equipment Code",
            "type": "string"
          },
          "equipment_id": {
            "format": "uuid",
            "title": "Equipment Id",
            "type": "string"
          },
          "equipment_name": {
            "title": "Equipment Name",
            "type": "string"
          },
          "installed_at": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Installed At"
          },
          "last_known_state": {
            "title": "Last Known State",
            "type": "string"
          },
          "last_state_updated_at": {
            "format": "date-time",
            "title": "Last State Updated At",
            "type": "string"
          },
          "manufacturer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manufacturer"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "serial_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Serial Number"
          }
        },
        "required": [
          "equipment_id",
          "cell_id",
          "equipment_code",
          "equipment_name",
          "equipment_class",
          "last_known_state",
          "last_state_updated_at"
        ],
        "title": "EquipmentDetail",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MaintenanceLogCreate": {
        "properties": {
          "equipment_id": {
            "format": "uuid",
            "title": "Equipment Id",
            "type": "string"
          },
          "failure_code": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Code"
          },
          "free_text_observation": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Free Text Observation"
          },
          "operator_state": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Operator State",
            "type": "string"
          },
          "reported_at": {
            "format": "date-time",
            "title": "Reported At",
            "type": "string"
          }
        },
        "required": [
          "equipment_id",
          "reported_at",
          "operator_state"
        ],
        "title": "MaintenanceLogCreate",
        "type": "object"
      },
      "MaintenanceLogItem": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "equipment_id": {
            "format": "uuid",
            "title": "Equipment Id",
            "type": "string"
          },
          "failure_code": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Code"
          },
          "free_text_observation": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Free Text Observation"
          },
          "maintenance_log_id": {
            "format": "uuid",
            "title": "Maintenance Log Id",
            "type": "string"
          },
          "operator_state": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Operator State",
            "type": "string"
          },
          "reported_at": {
            "format": "date-time",
            "title": "Reported At",
            "type": "string"
          }
        },
        "required": [
          "equipment_id",
          "reported_at",
          "operator_state",
          "maintenance_log_id",
          "created_at"
        ],
        "title": "MaintenanceLogItem",
        "type": "object"
      },
      "MaintenanceLogResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MaintenanceLogItem"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "MaintenanceLogResponse",
        "type": "object"
      },
...
      "SensorObservationCreate": {
        "properties": {
          "equipment_id": {
            "format": "uuid",
            "title": "Equipment Id",
            "type": "string"
          },
          "load_pct": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Load Pct"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "operating_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operating Hours"
          },
          "pressure_bar": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pressure Bar"
          },
          "rpm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rpm"
          },
          "sensor_quality": {
            "default": "unknown",
            "maxLength": 80,
            "minLength": 1,
            "title": "Sensor Quality",
            "type": "string"
          },
          "temperature_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature C"
          },
          "vibration_mm_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vibration Mm S"
          }
        },
        "required": [
          "equipment_id",
          "observed_at"
        ],
        "title": "SensorObservationCreate",
        "type": "object"
      },
      "SensorObservationItem": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "equipment_id": {
            "format": "uuid",
            "title": "Equipment Id",
            "type": "string"
          },
          "load_pct": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Load Pct"
          },
          "observation_id": {
            "format": "uuid",
            "title": "Observation Id",
            "type": "string"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "operating_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operating Hours"
          },
          "pressure_bar": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pressure Bar"
          },
          "rpm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rpm"
          },
          "sensor_quality": {
            "default": "unknown",
            "maxLength": 80,
            "minLength": 1,
            "title": "Sensor Quality",
            "type": "string"
          },
          "temperature_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature C"
          },
          "vibration_mm_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vibration Mm S"
          }
        },
        "required": [
          "equipment_id",
          "observed_at",
          "observation_id",
          "created_at"
        ],
        "title": "SensorObservationItem",
        "type": "object"
      },
      "SensorObservationResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SensorObservationItem"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "SensorObservationResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBasic": {
        "scheme": "basic",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API Core para MERO, motor de evaluacion de riesgo industrial.",
    "title": "MERO API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/assets/hierarchy": {
      "get": {
        "operationId": "list_asset_hierarchy_api_v1_assets_hierarchy_get",
        "parameters": [
          {
            "in": "query",
            "name": "tenant_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetHierarchyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "List Asset Hierarchy",
        "tags": [
          "assets"
        ]
      }
    },
    "/api/v1/discrepancies/audit/{observation_id}": {
      "post": {
        "operationId": "audit_observation_discrepancies_api_v1_discrepancies_audit__observation_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "observation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Observation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscrepancyAuditResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "Audit Observation Discrepancies",
        "tags": [
          "discrepancies"
        ]
      }
    },
    "/api/v1/equipment/{equipment_id}": {
      "get": {
        "operationId": "get_equipment_api_v1_equipment__equipment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "equipment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Equipment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquipmentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "Get Equipment",
        "tags": [
          "assets"
        ]
      }
    },
...
    "/api/v1/maintenance-logs": {
      "get": {
        "operationId": "list_maintenance_logs_api_v1_maintenance_logs_get",
        "parameters": [
          {
            "in": "query",
            "name": "equipment_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Equipment Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceLogResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "List Maintenance Logs",
        "tags": [
          "maintenance"
        ]
      },
      "post": {
        "operationId": "create_maintenance_log_api_v1_maintenance_logs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaintenanceLogCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceLogItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "Create Maintenance Log",
        "tags": [
          "maintenance"
        ]
      }
    },
...
    "/api/v1/sensor-observations": {
      "get": {
        "operationId": "list_sensor_observations_api_v1_sensor_observations_get",
        "parameters": [
          {
            "in": "query",
            "name": "equipment_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Equipment Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SensorObservationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "List Sensor Observations",
        "tags": [
          "sensor-observations"
        ]
      },
      "post": {
        "operationId": "create_sensor_observation_api_v1_sensor_observations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SensorObservationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SensorObservationItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "Create Sensor Observation",
        "tags": [
          "sensor-observations"
        ]
      }
    },
...
  }
}